refactor: 컨벤션 및 일부 메서드 리팩토링
