+++ src/js/components/pocket.js
@@ -1,6 +1,6 @@
 import { getQuotient } from "../util.js"
 import { TICKET } from "../constants/constant.js"
 
-export const getCount = (price) => {
+export const getTicketsCount = (price) => {
   return getQuotient(price, TICKET.PRICE)
 }
