+++ src/js/components/ticket.js
@@ -1,19 +1,10 @@
-import { TICKET } from "../constants/constant.js"
-import { generateRandomNumber, sortByNumber } from "../util.js"
+import { TICKET } from "./constants/constant.js"
+import { generateRandomNumber, sortByNumber } from "./util.js"
 
 class Ticket {
   #numbers
-
   constructor() {
     this.#numbers = new Set()
-  }
-
-  generateRandomNumbers() {
-    while (this.#numbers.size !== TICKET.SIZE) {
-      this.#numbers.add(generateRandomNumber(TICKET.MIN_NUM, TICKET.MAX_NUM))
-    }
-
-    this.#sortByNumber()
   }
 
   #sortByNumber() {
@@ -22,6 +13,14 @@ class Ticket {
 
   get numbers() {
     return [...this.#numbers]
+  }
+
+  generateRandomNumbers() {
+    while (this.#numbers.size !== TICKET.SIZE) {
+      this.#numbers.add(generateRandomNumber(TICKET.MIN_NUM, TICKET.MAX_NUM))
+    }
+
+    this.#sortByNumber()
   }
 }
