+++ src/js/models/lotto_model.js
@@ -1,3 +1,4 @@
+import Ticket from "../ticket.js"
 import { RANK } from "../constants/constant.js"
 import { getProfitRate } from "../util.js"
 
@@ -21,23 +22,11 @@ class LottoModel {
     }
   }
 
-  get tickets() {
-    return this.#tickets
-  }
-
-  get answerLotto() {
-    return this.#answerLotto
-  }
-
-  get lottoResult() {
-    return this.#lottoResult
-  }
-
-  addTicket(ticket) {
+  #addTicket(ticket) {
     this.#tickets.push(ticket)
   }
 
-  addAnswerLotto(numbers, bonus) {
+  #addAnswerLotto(numbers, bonus) {
     this.#answerLotto = { numbers, bonus }
   }
 
@@ -50,7 +39,31 @@ class LottoModel {
     return { match, bonusMatch }
   }
 
-  calculateLottosResult() {
+  #resetLottoResult() {
+    for (let i in this.#lottoResult) {
+      this.#lottoResult[i].count = 0
+    }
+  }
+
+  get tickets() {
+    return this.#tickets
+  }
+
+  get answerLotto() {
+    return this.#answerLotto
+  }
+
+  get lottoResult() {
+    return this.#lottoResult
+  }
+
+  generateRandomTicket() {
+    const ticket = new Ticket()
+    ticket.generateRandomNumbers()
+    this.#addTicket(ticket)
+  }
+
+  calculateLottosResult(numbers, bonus) {
     const calculateLottoResult = (ticket) => {
       const { match, bonusMatch } = this.#calculateMatch(ticket)
 
@@ -77,6 +90,7 @@ class LottoModel {
       key && this.#lottoResult[key].count++
     }
 
+    this.#addAnswerLotto(numbers, bonus)
     this.#tickets.forEach(calculateLottoResult)
   }
 
@@ -86,12 +100,6 @@ class LottoModel {
     }, 0)
 
     return getProfitRate(income, this.#tickets.length * 1000)
-  }
-
-  #resetLottoResult() {
-    for (let i in this.#lottoResult) {
-      this.#lottoResult[i].count = 0
-    }
   }
 
   init() {
